mnist/requirements.txt
pytorch-lightning
pillow
torchvision
hydra-core==1.2.0